// frontend/src/lib/proto/audit/v1/audit_pb.d.ts
export declare type AuditLog = Message<"audit.v1.AuditLog"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: string company_id = 2;
   */
  companyId: string;

  /**
   * 可空
   *
   * @generated from field: string department_id = 3;
   */
  departmentId: string;

  /**
   * @generated from field: string user_id = 4;
   */
  userId: string;

  /**
   * 操作者顯示名稱(join users)
   *
   * @generated from field: string user_name = 5;
   */
  userName: string;

  /**
   * create|update|delete|login|logout|print|force_logout|role_change|dispatch_cancel|void
   *
   * @generated from field: string action = 6;
   */
  action: string;

  /**
   * @generated from field: string resource_type = 7;
   */
  resourceType: string;

  /**
   * 可空
   *
   * @generated from field: string resource_id = 8;
   */
  resourceId: string;

  /**
   * JSON 字串(可空)
   *
   * @generated from field: string before_snapshot = 9;
   */
  beforeSnapshot: string;

  /**
   * JSON 字串(可空)
   *
   * @generated from field: string after_snapshot = 10;
   */
  afterSnapshot: string;

  /**
   * @generated from field: string ip_address = 11;
   */
  ipAddress: string;

  /**
   * @generated from field: string user_agent = 12;
   */
  userAgent: string;

  /**
   * RFC3339
   *
   * @generated from field: string created_at = 13;
   */
  createdAt: string;
};